+++ src/pages/employer/EmployerMyPage.tsx
@@ -1,11 +1,13 @@
 import { useEffect, useState } from "react";
 import type { ChangeEvent, JSX } from "react";
-import { useNavigate } from "react-router-dom";
+import { useNavigate, useLocation } from "react-router-dom";
 import { FaCamera, FaUser } from "react-icons/fa";
 import Swal from "sweetalert2";
 import "../../styles/employerMyPage.css";
-import { logout } from "../../api/authApi";
-import { getMyInfo, updateMyInfo, deleteMyAccount } from "../../api/commonApi";
+import "../../styles/employerMyPageReceive.css";
+import { getMyInfo } from "../../api/commonApi";
+import ProfileTab from "../../components/employer/EmployerMyPage/ProfileTab";
+import ReceivedRequestsTab from "../../components/employer/EmployerMyPage/ReceivedRequestsTab";
 
 interface User {
   id?: number;
@@ -14,21 +16,21 @@ interface User {
   profileImageUrl?: string | null;
 }
 
-type EditableField = "name" | "phone";
+type TabType = "profile" | "requests";
 
 export default function EmployerMyPage(): JSX.Element {
+  const navigate = useNavigate();
+  const location = useLocation();
+
+  // URL 경로에 따라 탭 결정
+  const currentTab: TabType = location.pathname.includes("employer-mypage-receive")
+    ? "requests"
+    : "profile";
+
   const [user, setUser] = useState<User | null>(null);
-  const [editableSections, setEditableSections] = useState<{ name: boolean; phone: boolean }>(
-    {
-      name: false,
-      phone: false,
-    }
-  );
   const [profileImage, setProfileImage] = useState<string | null>(null);
   const [loading, setLoading] = useState<boolean>(true);
-  const navigate = useNavigate();
 
-  // 사용자 정보 로드
   useEffect(() => {
     const fetchData = async (): Promise<void> => {
       try {
@@ -47,38 +49,14 @@ export default function EmployerMyPage(): JSX.Element {
     fetchData();
   }, []);
 
-  // 입력 필드 변경 핸들러
-  const handleChange = (field: EditableField, value: string): void => {
-    setUser((prev) => (prev ? { ...prev, [field]: value } : prev));
-  };
-
-  const handleSave = async (field: EditableField): Promise<void> => {
-    if (!user) return;
-
-    try {
-      await updateMyInfo({ [field]: user[field] ?? "" });
-      Swal.fire("완료", field === "name" ? "이름이 수정되었습니다." : "전화번호가 수정되었습니다.", "success");
-    } catch (error) {
-      const message = error instanceof Error ? error.message : "정보 수정 중 오류가 발생했습니다.";
-      Swal.fire("수정 실패", message, "error");
+  const handleTabClick = (tab: TabType): void => {
+    if (tab === "profile") {
+      navigate("/employer/employer-mypage");
+    } else {
+      navigate("/employer/employer-mypage-receive");
     }
   };
 
-  // 수정 모드 토글
-  const toggleEdit = (field: EditableField): void => {
-    setEditableSections((prev) => {
-      const wasEditing = prev[field];
-      const nextState = { ...prev, [field]: !prev[field] };
-
-      if (wasEditing) {
-        void handleSave(field);
-      }
-
-      return nextState;
-    });
-  };
-
-  // TODO: 프로필 이미지 변경 핸들러 (미구현)
   const handleProfileImageChange = async (e: ChangeEvent<HTMLInputElement>): Promise<void> => {
     const file = e.target.files?.[0];
     if (!file) return;
@@ -87,36 +65,8 @@ export default function EmployerMyPage(): JSX.Element {
     e.target.value = "";
   };
 
-  const handleWithdraw = async (): Promise<void> => {
-    const result = await Swal.fire({
-      icon: "warning",
-      title: "회원 탈퇴 하시겠습니까?",
-      text: "탈퇴 시 모든 정보가 삭제되며 복구할 수 없습니다.",
-      showCancelButton: true,
-      confirmButtonText: "탈퇴",
-      cancelButtonText: "취소",
-      confirmButtonColor: "var(--color-red)",
-    });
-
-    if (result.isConfirmed) {
-      try {
-        await deleteMyAccount();
-
-        Swal.fire("탈퇴 완료", "회원 탈퇴가 완료되었습니다.", "success");
-
-        // 로그아웃 처리
-        await logout();
-        localStorage.removeItem("token");
-        navigate("/");
-      } catch (error) {
-        const message = error instanceof Error ? error.message : "회원 탈퇴 중 오류가 발생했습니다.";
-        Swal.fire("탈퇴 실패", message, "error");
-      }
-    }
-  };
-
-  const handleNavClick = (path: string): void => {
-    navigate(path);
+  const handleUserUpdate = (updatedUser: User): void => {
+    setUser(updatedUser);
   };
 
   if (loading) {
@@ -164,8 +114,8 @@ export default function EmployerMyPage(): JSX.Element {
             <li>
               <button
                 type="button"
-                className="mypage-nav-checked"
-                onClick={() => handleNavClick("/employer/employer-mypage")}
+                className={currentTab === "profile" ? "mypage-nav-checked" : "mypage-nav-li"}
+                onClick={() => handleTabClick("profile")}
               >
                 내 프로필 수정
               </button>
@@ -173,50 +123,20 @@ export default function EmployerMyPage(): JSX.Element {
             <li>
               <button
                 type="button"
-                className="mypage-nav-li"
-                onClick={() => handleNavClick("/employer/employer-mypage-receive")}
+                className={currentTab === "requests" ? "mypage-nav-checked" : "mypage-nav-li"}
+                onClick={() => handleTabClick("requests")}
               >
                 받은 근무 요청
               </button>
             </li>
           </ul>
         </nav>
-        <div className="mypage-container">
-          <h1 className="mypage-title">기본정보</h1>
-          <div className="mypage-basic-info">
-            <div className="mypage-name">
-              <span className="mypage-label">이름</span>
-              <input
-                type="text"
-                value={user.name || ""}
-                disabled={!editableSections.name}
-                onChange={(e) => handleChange("name", e.target.value)}
-              />
-            </div>
-            <button className="mypage-edit-button" onClick={() => toggleEdit("name")}>
-              {editableSections.name ? "완료" : "수정"}
-            </button>
-          </div>
-          <hr />
-          <div className="mypage-phone">
-            <span className="mypage-label">전화번호</span>
-            <input
-              type="tel"
-              value={user.phone ?? ""}
-              disabled={!editableSections.phone}
-              onChange={(e) => handleChange("phone", e.target.value)}
-              placeholder="010-1234-5678"
-            />
-            <button className="mypage-edit-button" onClick={() => toggleEdit("phone")}>
-              {editableSections.phone ? "완료" : "수정"}
-            </button>
-          </div>
-          <div className="mypage-withdraw-section">
-            <button className="mypage-withdraw-button" onClick={handleWithdraw}>
-              회원 탈퇴 &gt;
-            </button>
-          </div>
-        </div>
+
+        {currentTab === "profile" && (
+          <ProfileTab user={user} onUserUpdate={handleUserUpdate} />
+        )}
+
+        {currentTab === "requests" && <ReceivedRequestsTab />}
       </div>
     </div>
   );
